agent-mesh-tools/llamaindex/requirements.txt
fastmcp
llama-index-core
llama-index-llms-ollama